+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ### HEAD
 
 _A description of your awesome changes here!_
+
+Features:
+
+- Raises error on `502 Bad Gateway` [#94](https://github.com/deliveroo/routemaster-drain/pull/94)
+
 ### 3.6.7 (2021-02-11)
 
 Features: